Refractor: Split Advanced Panel into components
- temporarily slightly increased window size
