+++ src-tauri/src/settings/mod.rs
@@ -55,7 +55,6 @@ pub struct ClickerSettings {
     pub custom_stop_zone_height: i32,
     pub disable_screenshots: bool,
     pub advanced_settings_enabled: bool,
-    pub explanation_mode: String,
     pub last_panel: String,
     pub show_stop_reason: bool,
     pub show_stop_overlay: bool,
@@ -108,7 +107,6 @@ impl Default for ClickerSettings {
             custom_stop_zone_height: 100,
             disable_screenshots: false,
             advanced_settings_enabled: true,
-            explanation_mode: "text".to_string(),
             last_panel: "simple".to_string(),
             show_stop_reason: true,
             show_stop_overlay: true,
